[SPARK-51701][PYTHON][TESTS] Move test objects to a separate file
Moves test objects to a separate file.

Some classes for tests should be placed in a clean separate file to avoid unnecessary dependencies.

No, test only.

The existing tests should pass.

No.

Closes #50503 from ueshin/issues/SPARK-51701/test_objects.

Authored-by: Takuya Ueshin <ueshin@databricks.com>
(cherry picked from commit 295d37fad3b67ac0c73629d5eaebb3baefaeea7e)
